feat: Allow fractional meal counts and add report sharing
This commit enhances the mess management system by enabling the use of fractional meal counts. Previously, meal counts were restricted to integers, which limited flexibility for recording partial meals.

Changes include:
- Updated the `count` property in the `Meal` model and `totalMeals` in `ReportData` from `int` to `double` to support decimal values.
- Modified data parsing logic (`Meal.fromMap`) and initial value assignments to correctly handle `double` types for meal counts.
- Adjusted the Google Apps Script to format meal-related columns in Google Sheets with one decimal place (`0.0`), ensuring accurate representation of fractional meals.
- Added a new feature to share the final mess report as an image, improving communication and usability.

+++ lib/models/mess_manager_models.dart
@@ -22,7 +22,7 @@ class Member {
 
 class Meal {
   final String memberId;
-  int count;
+  double count;
 
   Meal({required this.memberId, required this.count});
 
@@ -33,7 +33,7 @@ class Meal {
 
   factory Meal.fromMap(Map<String, dynamic> map) => Meal(
         memberId: map['memberId'] as String,
-        count: (map['count'] as num).toInt(),
+        count: (map['count'] as num).toDouble(),
       );
 }
 
@@ -156,7 +156,7 @@ class MemberExpense {
 class ReportData {
   final String memberId;
   final String memberName;
-  final int totalMeals;
+  final double totalMeals;
   final double initialDeposit; // Money given to manager
   final double personalExpense; // Money spent on market purchases
   final double mealCost; // Total cost of meals consumed (meals * mealRate)
